+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Implement email body decoding in the "Ping Details" dialog
 - Add a "Subject" field in the "Ping Details" dialog
 - Improve HTML email display in the "Ping Details" dialog
+- Add a link to check's details page in Slack notifications
 
 ## Bug Fixes
 - Fix downtime summary to handle months when the check didn't exist yet (#472)